requirements.txt
Django==4.2.3
mysqlclient==2.2.0
Pillow==10.0.0
Whoosh==2.7.4
django-haystack==3.2.1

